move policy distributions to new nets file
